feat(state): Wave C tip+apply satoshi integer cutover
Arm tip encoding v2 (b_satoshi) behind ceremony on fresh prod mesh, make
native Rocks tip/apply fee math satoshi-first, and fix worker-thread tip
ContextVar drift so wire probes stay state_consistent with ready×3.

+++ core/components/ports.py
@@ -25,7 +25,7 @@ class TxValidationResult:
 @dataclass
 class ApplyBlockResult:
     success: bool
-    burned: float = 0.0
+    burned: int = 0  # satoshi (Wave C); display via from_satoshi_float at edges
     state_root: str = ""
     error: str = ""
     native_applied: bool = False
